requirements.txt
flask==2.3.2
flask-restful==0.3.9
boto3==1.26.137
python-dotenv==1.0.0
pytest==7.3.1
requests==2.28.2
pyjwt==2.7.0
cryptography==41.0.1